Images: Add alt text and real sizes values to featured images
The alt text is not set on the image itself, it's set as post content, so it needs to be intentionally added to the featured image block.

The default sizes value assumes the image is full-width, which lead to loading the full-sized image for every photo on an archive page (100MB+ data). Setting the sizes to match the reality of the photo grid improves this by loading the medium or large image, improving the page size to around 5MB.