Attribution block: Restrict the attribution copy event to button clicks
Adding the copy event listener to `document` was catching all copy events, even copying other text on the page. This adds the event listener on button click, and removes it after the event fires.

Additionally, the copy listener function had to be moved to a separate function, because `callbacks.copyListener` changes with each state iteration, so `removeEventListener` never removed it.

Fixes #14